Fix: errori compilazione XInputJoystick + warning C4244 residui
1. FIX ERRORI DI COMPILAZIONE (XInputJoystick.cpp)
   - ERROR_DEVICE_NOT_CONNECTED: era ridefinito (warning C4005). Già
     definito in Windows SDK (winerror.h = 1167). Rimossa ridefinizione.
   - M_PI non dichiarato in MSVC: aggiunto #ifndef M_PI #define M_PI
     + #include <cmath> per sinf/cosf.
   - Aggiunto #pragma comment(lib, 'dinput8.lib') e 'dxguid.lib' per
     DirectInput nativo su MSVC.

2. FIX WARNING C4244 RESIDUI
   - Game.cpp: ~15 warning int->float e double->float
     * float = int_expr: aggiunto cast (float)
     * M_PI -> (float)M_PI in espressioni float
     * sf::Uint8 = float: aggiunto cast (sf::Uint8)
     * drawText con arg float: aggiunto cast (int)
   - Enemy.cpp L902/904: setPosition(... py - 2 + i*3) ->
     py - 2.f + (float)(i*3)
   - AudioManager.cpp L730: int = double (scale[] e' double) ->
     aggiunto cast (int)

Compila senza errori (g++ -std=c++17 -fsyntax-only -Wall -Wextra).

+++ src/AudioManager.cpp
@@ -727,7 +727,7 @@ void AudioManager::generateTrack(int trackIdx) {
     trackSamples.reserve(totalSamples);
 
     for(int bar = 0; bar < numBars; ++bar) {
-        int chordRoot = scale[prog[bar % 4]];
+        int chordRoot = (int)scale[prog[bar % 4]];
         bool isChorus = (bar >= 8 && bar <= 11) || (bar >= 20 && bar <= 23);
         // Boss: piu' intenso, chorus piu' frequente
         if (trackIdx == 4) isChorus = (bar % 4 == 0) || (bar % 4 == 1);
